requirements.txt
Flask
Werkzeug
opencv-python
scipy
pillow
pytesseract
ultralytics
vietocr
setuptools
numpy<2.0
pymongo[srv]
bcrypt
dotenv==0.9.9
appwrite
torch
torchvision
torchaudio